MarinePhysicsCore
Performance slightly improved

+++ src/SeaState.cpp
@@ -10,8 +10,8 @@ void SeaState::addSwell(const SwellConfig& config) {
     m_components.push_back(WaveComponent(config.height, config.period, config.phase, config.directionDeg));
 }
 
-double SeaState::getElevation(double x, double z, double t) const {
-    double eta = 0.0;
+float SeaState::getElevation(float x, float z, float t) const {
+    float eta = 0.0;
     for (const auto& w : m_components) {
         eta += w.getElevation(x, z, t);
     }
